fix(ci): Remove use of vars context (#9)
The vars context doesn't contain variables which
are opened against the main github repository when
the base branch is a fork. This causes GitHub PRs which
are made from forks to the main repository to fail.

This change removes the use of the vars context to
hard-code the variable instead.

This commit also makes two additional changes:

- adds a readme for the devcontainer
- Adds auto name detecting and scale parameters for different runner sizes